Zeitwerk compatibility; downgrade Spring
