+++ app/errors/not_authorized.rb
@@ -1,0 +1,1 @@
+class NotAuthorized < StandardError; end
