+++ app/errors/security_transgression.rb
@@ -1,0 +1,1 @@
+class SecurityTransgression < StandardError; end
