+++ app/errors/topic_not_found.rb
@@ -1,0 +1,1 @@
+class TopicNotFound < StandardError; end
