+++ config/initializers/extend_datetime.rb
@@ -1,2 +1,7 @@
 # Need to extend DateTime for millisecond timestamps
-require Rails.root.join('lib/date_time_extend')
+class DateTime
+  # Extend for millisecond timestamp
+  def floored_milliseconds_since_epoch
+    (to_f * 1000).floor
+  end
+end
