+++ config/initializers/v0_bindings_extensions.rb
