--- lib/errors.rb
@@ -1,3 +1,0 @@
-class SecurityTransgression < StandardError; end
-class TopicNotFound < StandardError; end
-class NotAuthorized < StandardError; end
